+++ Assets/Prefabs/CardGameView/Actions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71ebc61aa5d737445854436a9702c1c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
