+++ Assets/Prefabs/CardGameView/Actions/Card Action Buttons.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a9ff64913ad43c1428c7affb011349a2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
