+++ Assets/Prefabs/CardGameView/Viewer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7f68d23620554f42a51405eaba956aa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
