+++ Assets/Prefabs/CardGameView/Viewer/Playable Viewer.prefab.meta
