+++ Assets/Prefabs/CardGameView/Viewer/Stack Viewer.prefab.meta
