+++ Assets/Scripts/Cgs/CardGameView/Actions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7d8dcc7d01e67642af7bdb51c677fb7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
