+++ Assets/Scripts/Cgs/CardGameView/Actions/DieActions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa98b09e812317345a6ac97f13b8bb31
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
