+++ Assets/Scripts/Cgs/CardGameView/Actions/StackActions.cs
@@ -1,0 +1,12 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+using UnityEngine;
+
+namespace Cgs.CardGameView.Actions
+{
+    public class StackActions : MonoBehaviour
+    {
+    }
+}
