+++ Assets/Scripts/Cgs/CardGameView/Actions/StackActions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 585e7db9a4c6a5c49ae0098c8a0e1e26
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
