+++ Assets/Scripts/Cgs/CardGameView/Viewer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90de66742c0cbc14e81a5016bf3eb61f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
