+++ Assets/Scripts/Cgs/Cards/CardSelector.cs
@@ -5,6 +5,7 @@
 using System.Linq;
 using Cgs.CardGameView;
 using Cgs.CardGameView.Multiplayer;
+using Cgs.CardGameView.Viewer;
 using JetBrains.Annotations;
 using UnityEngine;
 using UnityEngine.EventSystems;
