+++ Assets/Scripts/Cgs/Cards/SearchResults.cs
@@ -5,7 +5,9 @@
 using System.Collections.Generic;
 using CardGameDef.Unity;
 using Cgs.CardGameView;
+using Cgs.CardGameView.Actions;
 using Cgs.CardGameView.Multiplayer;
+using Cgs.CardGameView.Viewer;
 using JetBrains.Annotations;
 using UnityEngine;
 using UnityEngine.UI;
