+++ Assets/Scripts/Cgs/Play/Drawer/CardDrawer.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using CardGameDef.Unity;
 using Cgs.CardGameView;
 using Cgs.CardGameView.Multiplayer;
+using Cgs.CardGameView.Viewer;
 using Cgs.Play.Multiplayer;
 using JetBrains.Annotations;
 using UnityEngine;
