+++ Assets/Scripts/Cgs/Play/PlayMatZoom.cs
@@ -4,6 +4,7 @@
 
 using System;
 using Cgs.CardGameView;
+using Cgs.CardGameView.Viewer;
 using JetBrains.Annotations;
 using UnityEngine;
 using UnityEngine.UI;
